// src/app/api/generated/model/persona-page-response.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PersonaResponse } from './persona-response';


/**
 * Pagina de personas del padron, segun los filtros pedidos
 */
export interface PersonaPageResponse { 
    /**
     * Personas de esta pagina
     */
    content?: Array<PersonaResponse>;
    /**
     * Numero de pagina devuelta, base cero
     */
    page?: number;
    /**
     * Cantidad de elementos por pagina
     */
    size?: number;
    /**
     * Cantidad total de personas que cumplen el filtro
     */
    totalElements?: number;
    /**
     * Cantidad total de paginas disponibles
     */
    totalPages?: number;
}